# Changing makeWorldspaceRay to return non-pointer.