--- a/include/camera.h
+++ b/include/camera.h
@@ -75,7 +75,7 @@ class Camera {
   Matrix4f* transformation();
 
   // Creates primary rays
-  Ray* makeWorldspaceRay(int i, int j, std::vector<float>*);
+  Ray makeWorldspaceRay(int i, int j, std::vector<float>*) const;
 };
 
 #endif