--- a/include/intersectables/instance.h
+++ b/include/intersectables/instance.h
@@ -43,7 +43,7 @@ class Instance : public Intersectable {
   Instance(Intersectable*, Matrix4f*);
   ~Instance();
 
-  HitRecord* intersect(Ray* ray) const;
+  HitRecord* intersect(const Ray& ray) const;
 };
 
 #endif