--- a/include/intersectables/plane.h
+++ b/include/intersectables/plane.h
@@ -31,7 +31,7 @@ class Plane : public Intersectable {
   // f(p) = dot(n, (p - a)) intersection:
   // f(p(t)) = 0. Solve for t.  Plug t_i into
   // p(t_i) will give intersection point
-  virtual HitRecord* intersect(Ray* ray) const;
+  virtual HitRecord* intersect(const Ray& ray) const;
 };
 
 #endif