--- a/include/intersectables/sphere.h
+++ b/include/intersectables/sphere.h
@@ -15,6 +15,6 @@ class Sphere : public Intersectable {
 
  public:
   Sphere(Material*, const Vector3f&, float);
-  virtual HitRecord* intersect(Ray* ray) const;
+  virtual HitRecord* intersect(const Ray& ray) const;
 };
 #endif