--- a/src/lights/pointLight.cpp
+++ b/src/lights/pointLight.cpp
@@ -7,4 +7,4 @@ PointLight::PointLight(Vector3f position, Spectrum* emission)
 
 HitRecord* PointLight::sample() const { return new HitRecord(new Vector3f(position), material); }
 
-HitRecord* PointLight::intersect(Ray*) const { return new HitRecord(); }
+HitRecord* PointLight::intersect(const Ray&) const { return new HitRecord(); }