Consequently changing all downstream usages to expect a const ref
instead.

I don't expect this to impact runtime heavily.