@simplay 2019-03-11
LGTM - many thanks  :smile_cat: :+1: 

FYI: I checked out your branch and ran some benchmarks on my machine. I also couldn't verify any noticeable performance impacts (see my measurements below). 

## before (simplay:master):

```
➜  ~/repos/cpptracer/build (master) ✔ ./benchmarks.x
2019-03-11 23:06:25
  Running ./benchmarks.x
Run on (12 X 4700 MHz CPU s)
  CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 256K (x6)
L3 Unified 12288K (x1)
  Load Average: 0.64, 0.99, 0.89
  ***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
  -------------------------------------------------------------------
  Benchmark                         Time             CPU   Iterations
  -------------------------------------------------------------------
  BM_TriangleIntersect            711 ns          711 ns       896024
  BM_SphereIntersect              560 ns          560 ns      1219065
  BM_ExplosionIntersect        166063 ns       166061 ns         4214
  BM_PlaneIntersect               461 ns          461 ns      1530754
  BM_PointLightIntersect          447 ns          447 ns      1560907
  BM_MeshTriangleTntersect        741 ns          741 ns       943835
  BM_Vector4fDot                 1.15 ns         1.15 ns    611749458
```

and
 
```
  ➜  ~/repos/cpptracer/build (master) ✔ hyperfine --warmup 2 --min-runs 2 "./main.x 300 300 7 1"
  Benchmark #1: ./main.x 300 300 7 1
  Time (mean ± σ):      4.160 s ±  0.071 s    [User: 47.149 s, System: 0.040 s]
                        Range (min … max):    4.110 s …  4.210 s
```

## after (panmari:const_refify_ray):

```
  ➜  ~/repos/cpptracer/build (dc3ffaf) ✔ ./benchmarks.x
  2019-03-11 23:09:50
  Running ./benchmarks.x
Run on (12 X 4700 MHz CPU s)
  CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 256K (x6)
L3 Unified 12288K (x1)
  Load Average: 1.44, 1.56, 1.16
  ***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
  -------------------------------------------------------------------
  Benchmark                         Time             CPU   Iterations
  -------------------------------------------------------------------
  BM_TriangleIntersect            702 ns          700 ns       906154
  BM_SphereIntersect              555 ns          553 ns      1260305
  BM_ExplosionIntersect        166257 ns       165642 ns         4162
  BM_PlaneIntersect               457 ns          456 ns      1542312
  BM_PointLightIntersect          450 ns          450 ns      1562580
  BM_MeshTriangleTntersect        705 ns          705 ns       986399
  BM_Vector4fDot                 1.15 ns         1.15 ns    606821771
```

and

```
  ➜  ~/repos/cpptracer/build (dc3ffaf) ✔ hyperfine --warmup 2 --min-runs 2 "./main.x 300 300 7 1"
  Benchmark #1: ./main.x 300 300 7 1
  Time (mean ± σ):      4.118 s ±  0.006 s    [User: 46.798 s, System: 0.036 s]
                        Range (min … max):    4.114 s …  4.122 s
```