@panmari 2019-03-12
Our current benchmarks only test "intersect" unfortunately. The only place I could see where this could matter is either "render" or "integrate".

We should add some benchmarks for these two.