@simplay 2019-03-12
I agree, we really should benchmark the integration and the renderer. For the time being, I always run hyperfine (https://github.com/sharkdp/hyperfine).